basic components added. Post Test OK but GUID ID will be added and general entities will be developed

+++ ShoppingMasterApp/ShoppingMasterApp.Application/Services/ProductService.cs
@@ -1,42 +1,56 @@
-﻿using ShoppingMasterApp.Application.CQRS.Commands.Product;
+﻿using AutoMapper;
+using ShoppingMasterApp.Application.CQRS.Commands.Category;
+using ShoppingMasterApp.Application.CQRS.Commands.Product;
 using ShoppingMasterApp.Application.Interfaces.Services;
 using ShoppingMasterApp.Domain.Entities;
-using ShoppingMasterApp.Domain.Exceptions;
 using ShoppingMasterApp.Domain.Interfaces.Repositories;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 
 namespace ShoppingMasterApp.Application.Services
 {
     public class ProductService : IProductService
     {
         private readonly IProductRepository _productRepository;
+        private readonly IMapper _mapper;
 
-        public ProductService(IProductRepository productRepository)
+        public ProductService(IProductRepository productRepository, IMapper mapper)
         {
             _productRepository = productRepository;
+            _mapper = mapper;
         }
 
-        public async Task<Product> CreateProductAsync(CreateProductCommand command)
+        public async Task<IEnumerable<Product>> GetAllProductsAsync()
         {
-            var product = new Product
-            {
-                Name = command.Name,
-                Price = command.Price,
-                Stock = command.Stock,
-                CategoryId = command.CategoryId
-            };
-
-            return await _productRepository.AddAsync(product);
+            return await _productRepository.GetAllAsync();
         }
 
         public async Task<Product> GetProductByIdAsync(int id)
         {
-            return await _productRepository.GetByIdAsync(id) ?? throw new ProductNotFoundException(id);
+            return await _productRepository.GetByIdAsync(id);
+        }
+
+        public async Task<Product> CreateProductAsync(CreateProductCommand command)
+        {
+            var product = _mapper.Map<Product>(command);
+            await _productRepository.AddAsync(product);
+            return product;
+        }
+
+        public async Task<Product> UpdateProductAsync(UpdateProductCommand command)
+        {
+            var product = _mapper.Map<Product>(command);
+            await _productRepository.UpdateAsync(product);
+            return product;
         }
 
         public async Task DeleteProductAsync(int id)
         {
-            var product = await GetProductByIdAsync(id);
-            await _productRepository.DeleteAsync(product);
+            var product = await _productRepository.GetByIdAsync(id);
+            if (product != null)
+            {
+                await _productRepository.DeleteAsync(product);
+            }
         }
     }
 }
